# src/swagger/swagger.json
{"swagger":"2.0","info":{"description":"Api para persistência e consulta de planetas do filme Star Wars.","version":"1.0","title":"Api-StarWars","termsOfService":"Termo de uso: Deve ser usada para estudos.","contact":{"name":"Augusto Leal","url":"http://www.augustoleal.com.br","email":"contato@augustoleal.com.br"},"license":{"name":"Licença - Open Source","url":"http://www.augustoleal.com.br"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"planet-controller","description":"Planet Controller"}],"paths":{"/starwars/planets":{"get":{"tags":["planet-controller"],"summary":"Retorna todos os planetas contidos no banco","description":"Essa operação retorna as informações de todos os planetas contidos no banco.","operationId":"getPlanetsUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"Retorna um PlanetServiceResponse com uma mensagem de sucesso e a lista de planetas","schema":{"$ref":"#/definitions/ServiceResponse"}},"401":{"description":"Unauthorized"},"500":{"description":"Caso tenhamos algum erro vamos retornar um PlanetServiceResponse com mensagem de erro interno","schema":{"$ref":"#/definitions/ServiceResponse"}},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/starwars/planets/delete/{id}":{"delete":{"tags":["planet-controller"],"summary":"Deleta um planeta baseado em seu id","description":"Essa operação deleta as informações de um planeta contido no banco.","operationId":"deletePlanetByIdUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"Retorna um ServiceResponse com uma mensagem de sucesso e a lista de planetas","schema":{"$ref":"#/definitions/ServiceResponse"}},"401":{"description":"Unauthorized"},"500":{"description":"Caso tenhamos algum erro vamos retornar um ServiceResponse com mensagem de erro interno","schema":{"$ref":"#/definitions/ServiceResponse"}},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/starwars/planets/id/{id}":{"get":{"tags":["planet-controller"],"summary":"Consulta um planeta baseado em seu id","description":"Essa operação retorna as informações de um planeta contido no banco.","operationId":"getPlanetByIdUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"Retorna um PlanetServiceResponse com uma mensagem de sucesso e a lista de planetas","schema":{"$ref":"#/definitions/PlanetServiceResponse"}},"401":{"description":"Unauthorized"},"500":{"description":"Caso tenhamos algum erro vamos retornar um PlanetServiceResponse com mensagem de erro interno","schema":{"$ref":"#/definitions/PlanetServiceResponse"}},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/starwars/planets/insert":{"post":{"tags":["planet-controller"],"summary":"Cadastrar um novo planeta","description":"Essa operação salva um novo registro com as informações do planeta.","operationId":"insertplanetUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"planet","description":"planet","required":true,"schema":{"$ref":"#/definitions/Planet"}}],"responses":{"200":{"description":"Retorna um ServiceResponse com uma mensagem de sucesso","schema":{"$ref":"#/definitions/ServiceResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"500":{"description":"Caso tenhamos algum erro vamos retornar um ServiceResponse com mensagem de erro interno","schema":{"$ref":"#/definitions/ServiceResponse"}},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/starwars/planets/nome/{nome}":{"get":{"tags":["planet-controller"],"summary":"Consulta um planeta baseado em seu nome","description":"Essa operação retorna as informações de um planeta contido no banco.","operationId":"getPlanetByNomeUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"nome","in":"path","description":"nome","required":true,"type":"string"}],"responses":{"200":{"description":"Retorna um PlanetServiceResponse com uma mensagem de sucesso e a lista de planetas","schema":{"$ref":"#/definitions/PlanetServiceResponse"}},"401":{"description":"Unauthorized"},"500":{"description":"Caso tenhamos algum erro vamos retornar um PlanetServiceResponse com mensagem de erro interno","schema":{"$ref":"#/definitions/PlanetServiceResponse"}},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"ServiceResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"success":{"type":"boolean"}}},"Planet":{"type":"object","properties":{"aparicoes":{"type":"integer","format":"int32"},"clima":{"type":"string"},"id":{"type":"string"},"nome":{"type":"string"},"terreno":{"type":"string"}}},"PlanetServiceResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"planet":{"$ref":"#/definitions/Planet"},"planetList":{"type":"array","items":{"$ref":"#/definitions/Planet"}},"success":{"type":"boolean"}}}}}